requirements.txt
python-dotenv
selenium 
beautifulsoup4 
requests
pytesseract 
Pillow
requests